run machine controller via leader election
this essentially means that we can have multiple instances and at the same time only one is operational.
before we acquire a lock we actually warm up caches mirroring the state of the API server.
machine controller will be run only if a proper lock was acquired.
note that the program terminates when the leadership was lost.