decreases verify tool timeout
the golang test runtime waits for individual subtests to complete before reporting the status.
if one of them is blocking/waiting and the global timeout is reached the status will not be reported/visible.

the global timeout is set to 20 minutes and the verify tool waits up to 60 hours for a machine to show up.
thus one faulty scenario prevents from showing the results for the whole group, which is confusing because it looks like all tests are broken.