Add the delete finalizer at the latest possible point
We currently add the delete finalizer to the machine at the very
beginning of the sync loop without even validating it before.

This moves the finalizer creation to happen right before the creation.

This will drastically reduce the occurence of situations where the
finalizer has to be manually deleted.

Related issue: #224