Move metric creation to controller
We currently have helper funcs to get a metrics struct inside cmd. This
poses a problem when trying to write tests for the controller, as the
tests will run into a NPE when the controller has no metrics struct but
it is not possible to use the existing helper func as that would result
in a cyclic dependency.

This commit also removes github.com/go-kit/kit/metrics as we have no
plans to implement a metric provider other than Prometheus.