Fix escaping in cloud-config
As quotation marks may be part of the password in the cloud config,
these must be escaped in order to let the quoted string be correct.
Also backslashes are interpreted as special characters and must be
escaped likewise.
For convenience every field of the cloud-config gets replaced,
regardless of whether or not its (currently) a valid character.