Fix conformance tests for CentOS (#573)
With this change, docker on CentOS now logs to json-file rather than
journald. This is required in order to make the conformance-tests pass on
machine-controller-generated nodes. Without it, a changing subset of the
tests constantly failed, their commonon denominator: They look for a
string in the output of container logs and get an empty log and no error
back.

Signed-off-by: Alvaro Aleman <alv2412@googlemail.com>