Use c1.small.x86 instance type on Packet (#812)
We've used t1.small.x86 instances for running E2E tests because we
don't need large instances. The t1.small.x86 instance type is first
generation, and as the first generation is deprecated, it's not
available anymore. The c1.small.x86 instances are bit more expensive,
but it's the smallest instance type currently available. Besides that,
we now run Packet E2E tests in the Amsterdam region.