Retry getting AWS instance (#886)
If machine-controller tries to get an instance before it's fully-created,
but after the instance request has been issued, it could
happen that it detects that there's no instance and create it again.
That could result in two or more instances created for one Machine object.
This happens more often in some AWS regions because some regions have
slower instance creation (e.g. us-east-1 and us-west-2).

Signed-off-by: Marko Mudrinić <mudrinic.mare@gmail.com>