Add enable-ingress-hostname option to cloudconfig
enable-ingress-hostname option will add a suffix to ingress ips to make
kubernetes believe that these are hostnames and will stop to add
iptables bypasses.

This is a workaround to be able to query a loadbalancer from within the
cluster when using proxy protocol.

This option will likely be removed in the future once a solution by
kubernetes has been released.