Fix RHEL subscription deregistration (#950)
The deregistration of a RHEL subscription failed if the machines fqdn
did not match it's node name. This happend for instance on OpenStack
clusters that add the .novalocal domain suffix.

This change sets the fqdn in the cloud-init userdata to the hostname
only to ensure that the machine name matches the registered name in the
RHSM or Satellite server.