Openstack support auth with projectName/ projectID (#1091)
* refactoring: store tenantName  and tenantID under ProjectName and ProjectID

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* openstack has deprecated OS_TENANT_NAME/OS_TENANT_ID in favor or OS_PROJECT_NAME/OS_PROJECT_ID, but these variables contain the same value.

The cloudProviderSpec has been updated to add projectName and projectID. We first try to read project variables. If they are not defined,
we fall back to tenant variables. Internally (ie Config struct), the values are stored under Project variables, even if the user has specified it
in tenant vars under cloudProviderSpec

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>

* apply gimps

Signed-off-by: Vincent Gramer <vincent@kubermatic.com>