rh_subscriprion bootstrap script (#1265)
* rh_subscriprion cloud-init

This Commit adds the rh_subscription cloud-init module to the yum
version of the bootstrap script.

Signed-off-by: Mattia Lavacca <lavacca.mattia@gmail.com>

* /etc/hostname patched also for rhel in Openstack

/etc/hostname is patched so that is contains the machineName not only
for Centos under Openstack, but also for RHEL.

Signed-off-by: Mattia Lavacca <lavacca.mattia@gmail.com>

* RHEL updated to 8.5 in Azure

Rhel has been updated to the 8.5 version. Beside this change, the part
of removing and adding the RHUI repo from yum isn't needed anymore, as
seems the problem has been solved by just updating the RHEL version,
thus it has been removed.

Signed-off-by: Mattia Lavacca <lavacca.mattia@gmail.com>

* yum update change in Azure

Signed-off-by: Mattia Lavacca <lavacca.mattia@gmail.com>

* update fixtures

Signed-off-by: Mattia Lavacca <lavacca.mattia@gmail.com>