Anexia: extend disk configuration
Adds a new config value "Disks", with each entry configuring the size
and performance type of a single disk. The config format can be used for
configuring multiple disks in the future, which is on our roadmap, but
right now only a single disk is supported - with the added feature of
defining the performance type.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>