Anexia: extend disk configuration (#1402)
* Anexia: cleanup Config/RawConfig attribute duplication

types.Config and types.RawConfig contained the same members, with
RawConfig using ConfigVars and Config storing the resolved data. This
commit does some cleanup to ease adding new config values in the future.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* Anexia: extend disk configuration

Adds a new config value "Disks", with each entry configuring the size
and performance type of a single disk. The config format can be used for
configuring multiple disks in the future, which is on our roadmap, but
right now only a single disk is supported - with the added feature of
defining the performance type.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* Anexia: update MachineDeployment example

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* CI: disable OperationSystemManager for Anexia

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>