Anexia: return reserved IP when instance IPs are requested
When creating a machine at, we always first reserve an IP address for
it. Later when retrieving the addresses of that machine, we rely on the
vminfo API, which has some delay as it retrieves the IPs from the
running VM.

This commit adds the reserved address to the list of addresses returned
from an instance, which should reduce provisioning time a bit and make
it more stable.

Also fixes a long-standing TODO comment: marking internal IPs as
internal.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>