[release/v1.45] Improvements for Anexia (#1405)
* Make anexia provider more resilient against errors (#1175)

* add conditions & tests

Original commit by @kstiehl, conflicts resolved by @LittleFox94, mainly
in pkg/cloudprovider/provider/anexia/provider.go

Signed-off-by: kstiehl <kevin.stiehl@numericas.de>
Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* Fix various linter errors

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* Add missing boilerplate to files

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* Anexia provider: add dummy updater for tests

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

Co-authored-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* docs(anexia): add info for templates (#1288)

* docs(anexia): add info for templates

Signed-off-by: Waleed Malik <ahmedwaleedmalik@gmail.com>

* Update anexia.md

Signed-off-by: Waleed Malik <ahmedwaleedmalik@gmail.com>

* anexia: increase HTTP client timeout to 120s (#1331)

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* Anexia: extend disk configuration (#1402)

* Anexia: cleanup Config/RawConfig attribute duplication

types.Config and types.RawConfig contained the same members, with
RawConfig using ConfigVars and Config storing the resolved data. This
commit does some cleanup to ease adding new config values in the future.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* Anexia: extend disk configuration

Adds a new config value "Disks", with each entry configuring the size
and performance type of a single disk. The config format can be used for
configuring multiple disks in the future, which is on our roadmap, but
right now only a single disk is supported - with the added feature of
defining the performance type.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* Anexia: update MachineDeployment example

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* CI: disable OperationSystemManager for Anexia

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

* Anexia: return reserved IP when instance IPs are requested

When creating a machine at, we always first reserve an IP address for
it. Later when retrieving the addresses of that machine, we rely on the
vminfo API, which has some delay as it retrieves the IPs from the
running VM.

This commit adds the reserved address to the list of addresses returned
from an instance, which should reduce provisioning time a bit and make
it more stable.

Also fixes a long-standing TODO comment: marking internal IPs as
internal.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>

Signed-off-by: Waleed Malik <ahmedwaleedmalik@gmail.com>
Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>
Co-authored-by: Kevin Stiehl <kevinstiehl@live.de>
Co-authored-by: Waleed Malik <ahmedwaleedmalik@gmail.com>