Adds field to check if a IP reservation is still valid, if the IP is no longer valid a new one is requested.
This fixes a bug where we use an IP that expired already for a VM,
if the VM provisioning failed, which then leads to a race condition.
In which we wait for the VM provisioning to finish, but it can't due the IP being no longer useable,
but we also never request a new IP in that case.

Signed-off-by: Zofia hagenguth <zhagenguth@anexia-it.com>