Anexia: various patches (request/response logging, CPU performance type, 404 fix on delete) (#1797)
* Anexia: really handle 404 response for DELETE

There already was a check for a 404 error being returned, but the client
library does not actually return 404 as error. Workaround is needed, as
that won't be fixed in go-anxcloud, as it's the legacy client - but
nicely commented and good workaround.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia.com>

* Anexia: use cloudproviderutil.HttpClient

This way we get request/response logging if required - even prefixed
with the Machine name, if applicable.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia.com>

* Anexia: allow to specify CPU performance type

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia.com>

---------

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia.com>
Co-authored-by: Mara Sophie Grosch <mgrosch@anexia.com>